Added function to recursively convert all markdown in a tree into html, then updated main.py to call it

+++ src/main.py
@@ -1,14 +1,17 @@
+import os
 import shutil
-from copy_static_content import copy_directory_contents, generate_page
+from copy_static_content import copy_directory_contents, generate_page, generate_pages_recursive
 
 
 
 def main():
     print(f"Clearing the public directory")
-    shutil.rmtree("public")
+    if os.path.exists("public"):
+        shutil.rmtree("public")
+    os.mkdir("public")
     copy_directory_contents("static", "public")
 
-    generate_page("content/index.md", "template.html", "public/index.html")
+    generate_pages_recursive("content", "template.html", "public")
 
 
 main()
